refactor: 💡 split MemberProfileTop to two components
